+++ packages/plugin-packages/mashroom-portal-demo-react-app2/src/js/indexApp.tsx
@@ -9,21 +9,23 @@ import type {Root} from 'react-dom/client';
 import type {MashroomPortalAppPluginBootstrapFunction} from '@mashroom/mashroom-portal/type-definitions';
 
 const bootstrap: MashroomPortalAppPluginBootstrapFunction = (portalAppHostElement, portalAppSetup, clientServices) => {
-    const {appConfig: {markdownMessage, pingButtonLabel}, user} = portalAppSetup;
+    const {serverSideRendered, appConfig: {markdownMessage, pingButtonLabel}, user} = portalAppSetup;
     const {messageBus, remoteLogger} = clientServices;
 
-    const ssrHost = portalAppHostElement.querySelector('[data-ssr-host="true"]');
-
     let root: Root;
-    if (ssrHost) {
+    if (serverSideRendered) {
         // SSR
         console.info('Hydrating React Demo App 2');
-        root = hydrateRoot(ssrHost, <App markdownMessage={markdownMessage} pingButtonLabel={pingButtonLabel} messageBus={messageBus}/>);
+        root = hydrateRoot(portalAppHostElement, (
+            <App markdownMessage={markdownMessage} pingButtonLabel={pingButtonLabel} messageBus={messageBus}/>
+        ));
     } else {
         // CSR
         console.info('Starting React Demo App 2');
         root = createRoot(portalAppHostElement);
-        root.render(<App markdownMessage={markdownMessage} pingButtonLabel={pingButtonLabel} messageBus={messageBus}/>);
+        root.render((
+            <App markdownMessage={markdownMessage} pingButtonLabel={pingButtonLabel} messageBus={messageBus}/>
+        ));
     }
 
     // Log on server